verzekeringsagent/requirements.txt
# PDF/document extraction
pymupdf>=1.24.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
python-docx>=1.1.0

# PII detection & anonymization
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# Encryption & database
cryptography>=42.0.0

# Claude API
anthropic>=0.40.0

# Environment
python-dotenv>=1.0.0
